requirements.txt
scapy
python-nmap
dnspython